Разобрали задачи на стримы, отрефакторили код

+++ src/main/java/ru/itpark/domain/Item/product/Product.java
@@ -11,11 +11,11 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class Product extends Item {
-    private int rating;
+    private double rating;
     private ProductType type;
     private LocalDateTime createdAt;
 
-    public Product(String name, int price, String description, int rating, ProductType type, LocalDateTime createdAt) {
+    public Product(String name, int price, String description, double rating, ProductType type, LocalDateTime createdAt) {
         super(name, price, description);
         this.rating = rating;
         this.type = type;
